requirements.txt
streamlit
pandas
python-docx
openpyxl
gspread
google-auth
